# project-management-ui/src/app/services/models/facture-request.ts
/* tslint:disable */
/* eslint-disable */
export interface FactureRequest {
  contactInformationRequestDto: string;
  contratIdRequestDto: number;
  coutRequestDto: number;
  descriptionRequestDto: string;
  endDateRequestDto: string;
  idRequestDto?: number;
  informationFacturationRequestDto: string;
  referenceRequestDto?: string;
  startDateRequestDto: string;
  statusRequestDto: string;
  termePaimentRequestDto: string;
  titreRequestDto: string;
}