# project-management-ui/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification - Ziedzaa",
    "description": "OpenApi documentation for Spring Security",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Ziedzaa",
      "url": "https://some-url.com",
      "email": "ziededucation@gmail.com"
    },
    "license": {
      "name": "Licence name",
      "url": "https://some-url.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:9090/api/v1",
      "description": "Local ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/factures": {
      "post": {
        "tags": [
          "Facture"
        ],
        "operationId": "createFacture",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FactureRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactureResponse"
                }
              }
            }
          }
        }
      }
    },
    "/contrats": {
      "post": {
        "tags": [
          "Contrat"
        ],
        "operationId": "createContrat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContratRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
...
    "/factures/{id}": {
      "get": {
        "tags": [
          "Facture"
        ],
        "operationId": "getFactureById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactureResponse"
                }
              }
            }
          }
        }
      }
    },
    "/factures/structure": {
      "get": {
        "tags": [
          "Facture"
        ],
        "operationId": "getFacturesByStructureId",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseFactureResponse"
                }
              }
            }
          }
        }
      }
    },
    "/factures/projet/{projetId}": {
      "get": {
        "tags": [
          "Facture"
        ],
        "operationId": "getFacturesByProjetId",
        "parameters": [
          {
            "name": "projetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseFactureResponse"
                }
              }
            }
          }
        }
      }
    },
    "/factures/contrat/{contratId}": {
      "get": {
        "tags": [
          "Facture"
        ],
        "operationId": "getFacturesByContratId",
        "parameters": [
          {
            "name": "contratId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseFactureResponse"
                }
              }
            }
          }
        }
      }
    },
    "/factures/appelOffre/{appelOffreId}": {
      "get": {
        "tags": [
          "Facture"
        ],
        "operationId": "getFacturesByAppelOffreId",
        "parameters": [
          {
            "name": "appelOffreId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseFactureResponse"
                }
              }
            }
          }
        }
      }
    },
    "/contrats/{id}": {
      "get": {
        "tags": [
          "Contrat"
        ],
        "operationId": "getContratById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContratResponse"
                }
              }
            }
          }
        }
      }
    },
    "/contrats/structure": {
      "get": {
        "tags": [
          "Contrat"
        ],
        "operationId": "getContratsByStructure",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseContratResponse"
                }
              }
            }
          }
        }
      }
    },
    "/contrats/projet/{projetId}": {
      "get": {
        "tags": [
          "Contrat"
        ],
        "operationId": "getContratsByProjetId",
        "parameters": [
          {
            "name": "projetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseContratResponse"
                }
              }
            }
          }
        }
      }
    },
    "/contrats/appelOffre/{appelOffreId}": {
      "get": {
        "tags": [
          "Contrat"
        ],
        "operationId": "getContratsByAppelOffreId",
        "parameters": [
          {
            "name": "appelOffreId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseContratResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FactureRequest": {
        "required": [
          "contactInformationRequestDto",
          "contratIdRequestDto",
          "coutRequestDto",
          "descriptionRequestDto",
          "endDateRequestDto",
          "informationFacturationRequestDto",
          "startDateRequestDto",
          "statusRequestDto",
          "termePaimentRequestDto",
          "titreRequestDto"
        ],
        "type": "object",
        "properties": {
          "idRequestDto": {
            "type": "integer",
            "format": "int32"
          },
          "referenceRequestDto": {
            "type": "string"
          },
          "titreRequestDto": {
            "maxLength": 30,
            "minLength": 2,
            "type": "string"
          },
          "descriptionRequestDto": {
            "type": "string"
          },
          "coutRequestDto": {
            "type": "number",
            "format": "double"
          },
          "startDateRequestDto": {
            "type": "string",
            "format": "date"
          },
          "endDateRequestDto": {
            "type": "string",
            "format": "date"
          },
          "statusRequestDto": {
            "type": "string"
          },
          "termePaimentRequestDto": {
            "type": "string"
          },
          "informationFacturationRequestDto": {
            "type": "string"
          },
          "contactInformationRequestDto": {
            "type": "string"
          },
          "contratIdRequestDto": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "FactureResponse": {
        "type": "object",
        "properties": {
          "idResponseDto": {
            "type": "integer",
            "format": "int32"
          },
          "referenceResponseDto": {
            "type": "string"
          },
          "titreResponseDto": {
            "type": "string"
          },
          "descriptionResponseDto": {
            "type": "string"
          },
          "coutResponseDto": {
            "type": "number",
            "format": "double"
          },
          "startDateResponseDto": {
            "type": "string",
            "format": "date"
          },
          "endDateResponseDto": {
            "type": "string",
            "format": "date"
          },
          "statusResponseDto": {
            "type": "string"
          },
          "termePaimentResponseDto": {
            "type": "string"
          },
          "informationFacturationResponseDto": {
            "type": "string"
          },
          "contactInformationResponseDto": {
            "type": "string"
          },
          "contratIdResponseDto": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ContratRequest": {
        "required": [
          "appelOffreIdRequestDto",
          "clauseDeResiliasionRequestDto",
          "coutRequestDto",
          "descriptionRequestDto",
          "endDateRequestDto",
          "livrablesRequestDto",
          "milestonesRequestDto",
          "startDateRequestDto",
          "termPaimentRequestDto",
          "titreRequestDto",
          "traveauxRequestDto"
        ],
        "type": "object",
        "properties": {
          "idRequestDto": {
            "type": "integer",
            "format": "int32"
          },
          "referenceRequestDto": {
            "type": "string"
          },
          "titreRequestDto": {
            "maxLength": 30,
            "minLength": 2,
            "type": "string"
          },
          "descriptionRequestDto": {
            "type": "string"
          },
          "coutRequestDto": {
            "type": "number",
            "format": "double"
          },
          "startDateRequestDto": {
            "type": "string",
            "format": "date"
          },
          "endDateRequestDto": {
            "type": "string",
            "format": "date"
          },
          "traveauxRequestDto": {
            "type": "string"
          },
          "livrablesRequestDto": {
            "type": "string"
          },
          "milestonesRequestDto": {
            "type": "string"
          },
          "termPaimentRequestDto": {
            "type": "string"
          },
          "clauseDeResiliasionRequestDto": {
            "type": "string"
          },
          "appelOffreIdRequestDto": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "PageResponseFactureResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FactureResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "ContratResponse": {
        "type": "object",
        "properties": {
          "idResponseDto": {
            "type": "integer",
            "format": "int32"
          },
          "referenceResponseDto": {
            "type": "string"
          },
          "titreResponseDto": {
            "type": "string"
          },
          "descriptionResponseDto": {
            "type": "string"
          },
          "coutResponseDto": {
            "type": "number",
            "format": "double"
          },
          "startDateResponseDto": {
            "type": "string",
            "format": "date"
          },
          "endDateResponseDto": {
            "type": "string",
            "format": "date"
          },
          "traveauxResponseDto": {
            "type": "string"
          },
          "livrablesResponseDto": {
            "type": "string"
          },
          "milestonesResponseDto": {
            "type": "string"
          },
          "termPaimentResponseDto": {
            "type": "string"
          },
          "clauseDeResiliasionResponseDto": {
            "type": "string"
          },
          "appelOffreIdResponseDto": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PageResponseContratResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContratResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}